Remove markdown from push notifications
