+++ Rocket.Chat.NotificationServiceExtension/NotificationService.swift
@@ -1,0 +1,37 @@
+//
+//  NotificationService.swift
+//  Rocket.Chat.NotificationServiceExtension
+//
+//  Created by Matheus Cardoso on 14/12/2018.
+//  Copyright © 2018 Rocket.Chat. All rights reserved.
+//
+
+import UserNotifications
+
+class NotificationService: UNNotificationServiceExtension {
+
+    var contentHandler: ((UNNotificationContent) -> Void)?
+    var bestAttemptContent: UNMutableNotificationContent?
+
+    override func didReceive(_ request: UNNotificationRequest, withContentHandler contentHandler: @escaping (UNNotificationContent) -> Void) {
+        self.contentHandler = contentHandler
+        bestAttemptContent = (request.content.mutableCopy() as? UNMutableNotificationContent)
+        
+        if let bestAttemptContent = bestAttemptContent {
+            // Modify the notification content here...
+            let attributedString = NSAttributedString(string: bestAttemptContent.body)
+            bestAttemptContent.body = MarkdownManager.shared.transformAttributedString(attributedString).string
+
+            contentHandler(bestAttemptContent)
+        }
+    }
+
+    override func serviceExtensionTimeWillExpire() {
+        // Called just before the extension will be terminated by the system.
+        // Use this as an opportunity to deliver your "best attempt" at modified content, otherwise the original push payload will be used.
+        if let contentHandler = contentHandler, let bestAttemptContent =  bestAttemptContent {
+            contentHandler(bestAttemptContent)
+        }
+    }
+
+}
