+++ Rocket.Chat/Managers/MarkdownManager.swift
@@ -104,6 +104,7 @@ fileprivate extension RCMarkdownParser {
         parser.strikeAttributes = [NSAttributedString.Key.strikethroughStyle.rawValue: NSNumber(value: NSUnderlineStyle.single.rawValue)]
         parser.linkAttributes = [NSAttributedString.Key.foregroundColor.rawValue: UIColor.darkGray]
 
+        #if !EXCLUDE_MARKDOWN_DOWNLOAD
         parser.downloadImage = { urlString, completion in
             guard let url = URL(string: urlString) else { return }
             guard let filename = DownloadManager.filenameFor(urlString) else { return }
@@ -127,6 +128,7 @@ fileprivate extension RCMarkdownParser {
                 }
             }
         }
+        #endif
 
         parser.headerAttributes = [
             1: [NSAttributedString.Key.font.rawValue: UIFont.boldSystemFont(ofSize: 26)],
